Adding functions to backend

+++ src/common.rs
@@ -1,11 +1,16 @@
-#[derive(Debug)]
+/// Represents a value that can appear in a spreadsheet.
+#[derive(Debug, Clone)]
+//Clone required for `Vec<Value>`(in graph.rs) to implement `Clone`
 pub enum Value {
-    Cell(i32, i32),
+    Cell(usize, usize),
     Const(isize),
     Oper(Box<Value>, Box<Value>, Operation), //value1 and value2, and the operation or command, respectively
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
+//Needed as Value has implemented a clone
+//Oper(Box<Value>, Box<Value>, Operation), //value1 and value2, and the operation or command, respectively
+//   |                                  ^^^^^^^^^ the trait `Clone` is not implemented for `Operation`
 pub enum Operation {
     Cons,
     Add,
